Reverted client refactoring
