+++ lib/occi/api/client.rb
@@ -1,4 +1,2 @@
-require 'occi/api/client_common'
-
 require 'occi/api/client_http'
 require 'occi/api/client_amqp'
